+++ amfora.go
@@ -11,7 +11,7 @@ import (
 )
 
 var (
-	version = "v1.7.2"
+	version = "v1.8.0"
 	commit  = "unknown"
 	builtBy = "unknown"
 )
@@ -52,14 +52,19 @@ func main() {
 
 	client.Init()
 
+	// Initialize lower-level cview app
+	if err = display.App.Init(); err != nil {
+		panic(err)
+	}
+
+	// Initialize Amfora's settings
 	display.Init(version, commit, builtBy)
 	display.NewTab()
-	display.NewTab() // Open extra tab and close it to fully initialize the app and wrapping
-	display.CloseTab()
 	if len(os.Args[1:]) > 0 {
 		display.URL(os.Args[1])
 	}
 
+	// Start
 	if err = display.App.Run(); err != nil {
 		panic(err)
 	}