+++ display/download.go
@@ -4,17 +4,20 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"mime"
 	"net/url"
 	"os"
+	"os/exec"
 	"path"
 	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/gdamore/tcell"
+	"github.com/gdamore/tcell/v2"
 	"github.com/makeworld-the-better-one/amfora/config"
 	"github.com/makeworld-the-better-one/amfora/structs"
+	"github.com/makeworld-the-better-one/amfora/sysopen"
 	"github.com/makeworld-the-better-one/go-gemini"
 	"github.com/schollz/progressbar/v3"
 	"github.com/spf13/viper"
@@ -22,8 +25,7 @@ import (
 )
 
 // For choosing between download and the portal - copy of YesNo basically
-var dlChoiceModal = cview.NewModal().
-	AddButtons([]string{"Download", "Open in portal", "Cancel"})
+var dlChoiceModal = cview.NewModal()
 
 // Channel to indicate what choice they made using the button text
 var dlChoiceCh = make(chan string)
@@ -31,108 +33,201 @@ var dlChoiceCh = make(chan string)
 var dlModal = cview.NewModal()
 
 func dlInit() {
+	panels.AddPanel("dl", dlModal, false, false)
+	panels.AddPanel("dlChoice", dlChoiceModal, false, false)
+
+	dlm := dlModal
+	chm := dlChoiceModal
 	if viper.GetBool("a-general.color") {
-		dlChoiceModal.SetButtonBackgroundColor(config.GetColor("btn_bg")).
-			SetButtonTextColor(config.GetColor("btn_text")).
-			SetBackgroundColor(config.GetColor("dl_choice_modal_bg")).
-			SetTextColor(config.GetColor("dl_choice_modal_text"))
-		dlChoiceModal.GetFrame().
-			SetBorderColor(config.GetColor("dl_choice_modal_text")).
-			SetTitleColor(config.GetColor("dl_choice_modal_text"))
+		chm.SetButtonBackgroundColor(config.GetColor("btn_bg"))
+		chm.SetButtonTextColor(config.GetColor("btn_text"))
+		chm.SetBackgroundColor(config.GetColor("dl_choice_modal_bg"))
+		chm.SetTextColor(config.GetColor("dl_choice_modal_text"))
+		form := chm.GetForm()
+		form.SetButtonBackgroundColorFocused(config.GetColor("btn_text"))
+		form.SetButtonTextColorFocused(config.GetColor("btn_bg"))
+		frame := chm.GetFrame()
+		frame.SetBorderColor(config.GetColor("dl_choice_modal_text"))
+		frame.SetTitleColor(config.GetColor("dl_choice_modal_text"))
 
-		dlModal.SetButtonBackgroundColor(config.GetColor("btn_bg")).
-			SetButtonTextColor(config.GetColor("btn_text")).
-			SetBackgroundColor(config.GetColor("dl_modal_bg")).
-			SetTextColor(config.GetColor("dl_modal_text"))
-		dlModal.GetFrame().
-			SetBorderColor(config.GetColor("dl_modal_text")).
-			SetTitleColor(config.GetColor("dl_modal_text"))
+		dlm.SetButtonBackgroundColor(config.GetColor("btn_bg"))
+		dlm.SetButtonTextColor(config.GetColor("btn_text"))
+		dlm.SetBackgroundColor(config.GetColor("dl_modal_bg"))
+		dlm.SetTextColor(config.GetColor("dl_modal_text"))
+		form = dlm.GetForm()
+		form.SetButtonBackgroundColorFocused(config.GetColor("btn_text"))
+		form.SetButtonTextColorFocused(config.GetColor("btn_bg"))
+		frame = dlm.GetFrame()
+		frame.SetBorderColor(config.GetColor("dl_modal_text"))
+		frame.SetTitleColor(config.GetColor("dl_modal_text"))
 	} else {
-		dlChoiceModal.SetButtonBackgroundColor(tcell.ColorWhite).
-			SetButtonTextColor(tcell.ColorBlack).
-			SetBackgroundColor(tcell.ColorBlack).
-			SetTextColor(tcell.ColorWhite)
-		dlChoiceModal.SetBorderColor(tcell.ColorWhite)
-		dlChoiceModal.GetFrame().SetTitleColor(tcell.ColorWhite)
+		chm.SetButtonBackgroundColor(tcell.ColorWhite)
+		chm.SetButtonTextColor(tcell.ColorBlack)
+		chm.SetBackgroundColor(tcell.ColorBlack)
+		chm.SetTextColor(tcell.ColorWhite)
+		chm.SetBorderColor(tcell.ColorWhite)
+		chm.GetFrame().SetTitleColor(tcell.ColorWhite)
+		form := chm.GetForm()
+		form.SetButtonBackgroundColorFocused(tcell.ColorBlack)
+		form.SetButtonTextColorFocused(tcell.ColorWhite)
 
-		dlModal.SetButtonBackgroundColor(tcell.ColorWhite).
-			SetButtonTextColor(tcell.ColorBlack).
-			SetBackgroundColor(tcell.ColorBlack).
-			SetTextColor(tcell.ColorWhite)
-		dlModal.GetFrame().
-			SetBorderColor(tcell.ColorWhite).
-			SetTitleColor(tcell.ColorWhite)
+		dlm.SetButtonBackgroundColor(tcell.ColorWhite)
+		dlm.SetButtonTextColor(tcell.ColorBlack)
+		dlm.SetBackgroundColor(tcell.ColorBlack)
+		dlm.SetTextColor(tcell.ColorWhite)
+		form = dlm.GetForm()
+		form.SetButtonBackgroundColorFocused(tcell.ColorBlack)
+		form.SetButtonTextColorFocused(tcell.ColorWhite)
+		frame := dlm.GetFrame()
+		frame.SetBorderColor(tcell.ColorWhite)
+		frame.SetTitleColor(tcell.ColorWhite)
 	}
 
-	dlChoiceModal.SetBorder(true)
-	dlChoiceModal.GetFrame().SetTitleAlign(cview.AlignCenter)
-	dlChoiceModal.SetDoneFunc(func(buttonIndex int, buttonLabel string) {
+	chm.AddButtons([]string{"Open", "Download", "Cancel"})
+	chm.SetBorder(true)
+	chm.GetFrame().SetTitleAlign(cview.AlignCenter)
+	chm.SetDoneFunc(func(buttonIndex int, buttonLabel string) {
 		dlChoiceCh <- buttonLabel
 	})
 
-	dlModal.SetBorder(true)
-	dlModal.GetFrame().
-		SetTitleAlign(cview.AlignCenter).
-		SetTitle(" Download ")
-	dlModal.SetDoneFunc(func(buttonIndex int, buttonLabel string) {
+	dlm.SetBorder(true)
+	frame := dlm.GetFrame()
+	frame.SetTitleAlign(cview.AlignCenter)
+	frame.SetTitle(" Download ")
+	dlm.SetDoneFunc(func(buttonIndex int, buttonLabel string) {
 		if buttonLabel == "Ok" {
-			tabPages.SwitchToPage(strconv.Itoa(curTab))
+			panels.HidePanel("dl")
 			App.SetFocus(tabs[curTab].view)
 			App.Draw()
 		}
 	})
+}
+
+func getMediaHandler(resp *gemini.Response) config.MediaHandler {
+	def := config.MediaHandler{
+		Cmd:      nil,
+		NoPrompt: false,
+		Stream:   false,
+	}
+
+	mediatype, _, err := mime.ParseMediaType(resp.Meta)
+	if err != nil {
+		return def
+	}
+
+	if ret, ok := config.MediaHandlers[mediatype]; ok {
+		return ret
+	}
+
+	splitType := strings.Split(mediatype, "/")[0]
+	if ret, ok := config.MediaHandlers[splitType]; ok {
+		return ret
+	}
+
+	if ret, ok := config.MediaHandlers["*"]; ok {
+		return ret
+	}
+
+	return def
 }
 
 // dlChoice displays the download choice modal and acts on the user's choice.
 // It should run in a goroutine.
 func dlChoice(text, u string, resp *gemini.Response) {
-	defer resp.Body.Close()
+	mediaHandler := getMediaHandler(resp)
+	var choice string
 
-	parsed, err := url.Parse(u)
-	if err != nil {
-		Error("URL Error", err.Error())
+	if mediaHandler.NoPrompt {
+		choice = "Open"
+	} else {
+		dlChoiceModal.SetText(text)
+		panels.ShowPanel("dlChoice")
+		panels.SendToFront("dlChoice")
+		App.SetFocus(dlChoiceModal)
+		App.Draw()
+		choice = <-dlChoiceCh
+	}
+
+	if choice == "Download" {
+		panels.HidePanel("dlChoice")
+		App.Draw()
+		downloadURL(config.DownloadsDir, u, resp)
+		resp.Body.Close() // Only close when the file is downloaded
+		return
+	}
+	if choice == "Open" {
+		panels.HidePanel("dlChoice")
+		App.Draw()
+		open(u, resp)
 		return
 	}
 
-	dlChoiceModal.SetText(text)
-	tabPages.ShowPage("dlChoice")
-	tabPages.SendToFront("dlChoice")
-	App.SetFocus(dlChoiceModal)
+	// They chose the "Cancel" button
+	panels.HidePanel("dlChoice")
+	App.SetFocus(tabs[curTab].view)
+	App.Draw()
+}
+
+// open performs the same actions as downloadURL except it also opens the file.
+// If there is no system viewer configured for the particular mediatype, it opens it
+// with the default system viewer.
+func open(u string, resp *gemini.Response) {
+	mediaHandler := getMediaHandler(resp)
+
+	if mediaHandler.Stream {
+		// Run command with downloaded data from stdin
+
+		cmd := mediaHandler.Cmd
+		var proc *exec.Cmd
+		if len(cmd) == 1 {
+			proc = exec.Command(cmd[0])
+		} else {
+			proc = exec.Command(cmd[0], cmd[1:]...)
+		}
+		proc.Stdin = resp.Body
+
+		err := proc.Start()
+		if err != nil {
+			Error("File Opening Error", "Error executing custom command: "+err.Error())
+			return
+		}
+		Info("Opened with " + cmd[0])
+		return
+	}
+
+	path := downloadURL(config.TempDownloadsDir, u, resp)
+	if path == "" {
+		return
+	}
+
+	panels.HidePanel("dl")
+	App.SetFocus(tabs[curTab].view)
 	App.Draw()
 
-	choice := <-dlChoiceCh
-	if choice == "Download" {
-		tabPages.HidePage("dlChoice")
-		App.Draw()
-		downloadURL(u, resp)
-		return
-	}
-	if choice == "Open in portal" {
-		// Open in mozz's proxy
-		portalURL := u
-		if parsed.RawQuery != "" {
-			// Remove query and add encoded version on the end
-			query := parsed.RawQuery
-			parsed.RawQuery = ""
-			portalURL = parsed.String() + "%3F" + query
+	if mediaHandler.Cmd == nil {
+		// Open with system default viewer
+		_, err := sysopen.Open(path)
+		if err != nil {
+			Error("System Viewer Error", err.Error())
+			return
 		}
-		portalURL = strings.TrimPrefix(portalURL, "gemini://") + "?raw=1"
-		ok := handleHTTP("https://portal.mozz.us/gemini/"+portalURL, false)
-		if ok {
-			tabPages.SwitchToPage(strconv.Itoa(curTab))
-			App.SetFocus(tabs[curTab].view)
-			App.Draw()
+		Info("Opened in default system viewer")
+	} else {
+		cmd := mediaHandler.Cmd
+		err := exec.Command(cmd[0], append(cmd[1:], path)...).Start()
+		if err != nil {
+			Error("File Opening Error", "Error executing custom command: "+err.Error())
+			return
 		}
-		return
+		Info("Opened with " + cmd[0])
 	}
-	tabPages.SwitchToPage(strconv.Itoa(curTab))
-	App.SetFocus(tabs[curTab].view)
 	App.Draw()
 }
 
 // downloadURL pulls up a modal to show download progress and saves the URL content.
 // downloadPage should be used for Page content.
-func downloadURL(u string, resp *gemini.Response) {
+// Returns location downloaded to or an empty string on error.
+func downloadURL(dir, u string, resp *gemini.Response) string {
 	_, _, width, _ := dlModal.GetInnerRect()
 	// Copy of progressbar.DefaultBytesSilent with custom width
 	bar := progressbar.NewOptions64(
@@ -146,15 +241,15 @@ func downloadURL(u string, resp *gemini.Response) {
 	)
 	bar.RenderBlank() //nolint:errcheck
 
-	savePath, err := downloadNameFromURL(u, "")
+	savePath, err := downloadNameFromURL(dir, u, "")
 	if err != nil {
 		Error("Download Error", "Error deciding on file name: "+err.Error())
-		return
+		return ""
 	}
 	f, err := os.OpenFile(savePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
 	if err != nil {
 		Error("Download Error", "Error creating download file: "+err.Error())
-		return
+		return ""
 	}
 	defer f.Close()
 
@@ -172,19 +267,19 @@ func downloadURL(u string, resp *gemini.Response) {
 	// Display
 	dlModal.ClearButtons()
 	dlModal.AddButtons([]string{"Downloading..."})
-	tabPages.ShowPage("dl")
-	tabPages.SendToFront("dl")
+	panels.ShowPanel("dl")
+	panels.SendToFront("dl")
 	App.SetFocus(dlModal)
 	App.Draw()
 
 	_, err = io.Copy(io.MultiWriter(f, bar), resp.Body)
 	done = true
 	if err != nil {
-		tabPages.HidePage("dl")
+		panels.HidePanel("dl")
 		Error("Download Error", err.Error())
 		f.Close()
 		os.Remove(savePath) // Remove partial file
-		return
+		return ""
 	}
 	dlModal.SetText(fmt.Sprintf("Download complete! File saved to %s.", savePath))
 	dlModal.ClearButtons()
@@ -192,6 +287,8 @@ func downloadURL(u string, resp *gemini.Response) {
 	dlModal.GetForm().SetFocus(100)
 	App.SetFocus(dlModal)
 	App.Draw()
+
+	return savePath
 }
 
 // downloadPage saves the passed Page to a file.
@@ -202,9 +299,9 @@ func downloadPage(p *structs.Page) (string, error) {
 	var err error
 
 	if p.Mediatype == structs.TextGemini {
-		savePath, err = downloadNameFromURL(p.URL, ".gmi")
+		savePath, err = downloadNameFromURL(config.DownloadsDir, p.URL, ".gmi")
 	} else {
-		savePath, err = downloadNameFromURL(p.URL, ".txt")
+		savePath, err = downloadNameFromURL(config.DownloadsDir, p.URL, ".txt")
 	}
 	if err != nil {
 		return "", err
@@ -221,13 +318,13 @@ func downloadPage(p *structs.Page) (string, error) {
 // downloadNameFromURL takes a URl and returns a safe download path that will not overwrite any existing file.
 // ext is an extension that will be added if the file has no extension, and for domain only URLs.
 // It should include the dot.
-func downloadNameFromURL(u string, ext string) (string, error) {
+func downloadNameFromURL(dir, u, ext string) (string, error) {
 	var name string
 	var err error
 	parsed, _ := url.Parse(u)
 	if parsed.Path == "" || path.Base(parsed.Path) == "/" {
 		// No file, just the root domain
-		name, err = getSafeDownloadName(parsed.Hostname()+ext, true, 0)
+		name, err = getSafeDownloadName(dir, parsed.Hostname()+ext, true, 0)
 		if err != nil {
 			return "", err
 		}
@@ -238,23 +335,23 @@ func downloadNameFromURL(u string, ext string) (string, error) {
 			// No extension
 			name += ext
 		}
-		name, err = getSafeDownloadName(name, false, 0)
+		name, err = getSafeDownloadName(dir, name, false, 0)
 		if err != nil {
 			return "", err
 		}
 	}
-	return filepath.Join(config.DownloadsDir, name), nil
+	return filepath.Join(dir, name), nil
 }
 
 // getSafeDownloadName is used by downloads.go only.
-// It returns a modified name that is unique for the downloads folder.
+// It returns a modified name that is unique for the specified folder.
 // This way duplicate saved files will not overwrite each other.
 //
 // lastDot should be set to true if the number added to the name should come before
 // the last dot in the filename instead of the first.
 //
 // n should be set to 0, it is used for recursiveness.
-func getSafeDownloadName(name string, lastDot bool, n int) (string, error) {
+func getSafeDownloadName(dir, name string, lastDot bool, n int) (string, error) {
 	// newName("test.txt", 3) -> "test(3).txt"
 	newName := func() string {
 		if n <= 0 {
@@ -271,7 +368,7 @@ func getSafeDownloadName(name string, lastDot bool, n int) (string, error) {
 		return name[:idx] + "(" + strconv.Itoa(n) + ")" + name[idx:]
 	}
 
-	d, err := os.Open(config.DownloadsDir)
+	d, err := os.Open(dir)
 	if err != nil {
 		return "", err
 	}
@@ -285,7 +382,7 @@ func getSafeDownloadName(name string, lastDot bool, n int) (string, error) {
 	for i := range files {
 		if nn == files[i] {
 			d.Close()
-			return getSafeDownloadName(name, lastDot, n+1)
+			return getSafeDownloadName(dir, name, lastDot, n+1)
 		}
 	}
 	d.Close()