+++ display/newtab.go
@@ -21,12 +21,13 @@ Happy browsing!
 
 => about:bookmarks Bookmarks
 => about:subscriptions Subscriptions
+=> about:about All internal pages
 
 ## Learn more about Amfora!
 
 => https://github.com/makeworld-the-better-one/amfora Amfora homepage
 => https://github.com/makeworld-the-better-one/amfora/wiki Amfora Wiki [GitHub]
-=> gemini://makeworld.gq/amfora-wiki/ Amfora Wiki [On Gemini!]
+=> gemini://makeworld.space/amfora-wiki/ Amfora Wiki [On Gemini!]
 
 => //gemini.circumlunar.space Project Gemini
 `